Fix Plugin Override Function Capabilities
In order to ensure that plugin configuration that is overridden with a
function instead of a table is still merged with the default plugin
configuration, this commit splits up the if/elseif/else statement in
lua/core/utils.lua.

This is necessary as we want to take the resulting table returned from
the override function and merge it with the default_table before setting
up the plugin.

Fixes [this issue](https://github.com/NvChad/NvChad/issues/1239)